backend/requirements-abs-py39.txt
aiosqlite==0.17.0
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
databases==0.4.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
PyJWT==2.1.0
python-decouple==3.4
python-dotenv==0.18.0
PyYAML==5.4.1
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1